requirements.txt
streamlit
google-generativeai
SpeechRecognition
Pillow
fpdf